+++ examples/vue/2.7-basic/src/Post.vue
@@ -20,7 +20,7 @@ export default defineComponent({
   emits: ['setPostId'],
   setup(props) {
     const { isPending, isError, isFetching, data, error } = useQuery({
-      queryKey: ['post', props.postId],
+      queryKey: ['post', () => props.postId],
       queryFn: () => fetcher(props.postId),
     })