+++ packages/query-core/src/__tests__/utils.test-d.tsx
@@ -1,4 +1,4 @@
-import { describe, expectTypeOf, it } from 'vitest'
+import { assertType, describe, expectTypeOf, it } from 'vitest'
 import { QueryClient } from '../queryClient'
 import type { QueryFilters } from '../utils'
 import type { DataTag, QueryKey } from '../types'
@@ -35,11 +35,36 @@ describe('QueryFilters', () => {
     }
 
     const queryClient = new QueryClient()
-
     const data = queryClient.getQueryData(a.queryKey!)
     expectTypeOf(data).toEqualTypeOf<unknown>()
 
     const error = queryClient.getQueryState(a.queryKey!)?.error
     expectTypeOf(error).toEqualTypeOf<Error | null | undefined>()
+  })
+
+  it('should allow a partial query key to be passed', () => {
+    const filters: QueryFilters<readonly ['key', { a: number; b: string }]> = {
+      queryKey: ['key'],
+    }
+
+    expectTypeOf(filters.queryKey).toEqualTypeOf<
+      | undefined
+      | readonly []
+      | ['key']
+      | readonly [
+          'key',
+          {
+            a: number
+            b: string
+          },
+        ]
+    >()
+  })
+
+  it('should error on invalid query keys', () => {
+    assertType<QueryFilters<readonly ['key', { a: number; b: string }]>>({
+      // @ts-expect-error cannot pass invalid query key
+      queryKey: ['invalid', { a: 1, b: '1' }],
+    })
   })
 })