+++ packages/svelte-query-persist-client/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @tanstack/svelte-query-persist-client
 
+## 6.0.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.7
+  - @tanstack/svelte-query@6.0.6
+
+## 6.0.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.6
+  - @tanstack/svelte-query@6.0.5
+
+## 6.0.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.5
+  - @tanstack/svelte-query@6.0.4
+
+## 6.0.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.4
+  - @tanstack/svelte-query@6.0.3
+
+## 6.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.3
+  - @tanstack/svelte-query@6.0.2
+
+## 6.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.2
+  - @tanstack/svelte-query@6.0.1
+
+## 6.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`846d53d`](https://github.com/TanStack/query/commit/846d53d98992d50606c40634efa43dea9965b787)]:
+  - @tanstack/query-persist-client-core@5.91.1
+
+## 6.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`5cd86c6`](https://github.com/TanStack/query/commit/5cd86c6ef1720b87b13e1ab70ee823616f1f029a)]:
+  - @tanstack/query-persist-client-core@5.91.0
+
 ## 6.0.0
 
 ### Major Changes